+++ CHANGELOG.md
@@ -47,6 +47,9 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 #### :hourglass: Performance
 #### :mortar_board: Walkthrough / Help
 #### :hammer: Development
+* Replace `sinon` with `vitest`'s built in spy/mock library ([#12058])
+
+[#12058]: https://github.com/openstreetmap/iD/pull/12058
 
 
 # 2.39.2